requirements.txt
Flask
Flask-Peewee
mysqlclient